Move up the dashboard pacakge one level
