+++ dolphin/async/src/main/java/edu/snu/cay/dolphin/async/dashboard/DashboardConfProvider.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.snu.cay.dolphin.async.metric.dashboard;
+package edu.snu.cay.dolphin.async.dashboard;
 
-import edu.snu.cay.dolphin.async.metric.dashboard.parameters.DashboardHostAddress;
+import edu.snu.cay.dolphin.async.dashboard.parameters.DashboardHostAddress;
 import org.apache.reef.tang.Configuration;
 import org.apache.reef.tang.JavaConfigurationBuilder;
 import org.apache.reef.tang.Tang;
