+++ dolphin/async/src/main/java/edu/snu/cay/dolphin/async/dashboard/DashboardLauncher.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.snu.cay.dolphin.async.metric.dashboard;
+package edu.snu.cay.dolphin.async.dashboard;
 
 import java.io.IOException;
 import java.net.ConnectException;
