+++ dolphin/async/src/main/java/edu/snu/cay/dolphin/async/dashboard/parameters/DashboardPort.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.snu.cay.dolphin.async.metric.dashboard.parameters;
+package edu.snu.cay.dolphin.async.dashboard.parameters;
 
 import org.apache.reef.tang.annotations.Name;
 import org.apache.reef.tang.annotations.NamedParameter;
